+++ CryptoLib/src/scard/scard_library.h
@@ -20,7 +20,7 @@ uint8_t sc_create_ctx(void);
 void sc_delete_ctx(void);
 uint8_t sc_get_available_readers(void);
 uint8_t sc_get_reader_status(void);
-uint8_t sc_apdu_transmit(uint8_t* cmd, uint32_t cmdLen, uint8_t* resp, uint32_t* respLen);
+uint8_t sc_apdu_transmit(uint8_t* cmd, uint32_t cmdLen, uint8_t* resp, uint64_t* respLen);
 uint8_t sc_card_connect(void);
 uint8_t sc_card_disconnect(void);
